refactor: LLM 서비스를 인프라 계층으로 분리 - LLM 관련 로직을 독립된 llm.service.ts로 분리 - common/types에 Question 타입 생성 - infra/formats 폴더 생성 및 QuestionResponseFormat 별도 파일로 분리 - QuestionResponseFormat에서 enum 객체 자체를 참조하는 문제 해결
