+++ src/common/types/question.type.ts
@@ -1,0 +1,7 @@
+import { QuestionType } from "src/common/enums/question-type.enum";
+
+export interface Question {
+    id: number;
+    question: string;
+    questionType: QuestionType;
+}
