+++ src/modules/master-portfolios/master-portfolios.service.ts
@@ -1,52 +1,21 @@
 import { Injectable, InternalServerErrorException } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
-import { QuestionType } from 'src/common/enums/question-type.enum';
-import { PromptLoader } from 'src/common/utils/prompt.loader';
 import { Questions } from './entities/questions.entity';
 import { Repository } from 'typeorm';
 import { MasterPortfolio } from './entities/master-portfolios.entity';
 import { MasterPortfolioNotFoundException } from 'src/common/exceptions/custom.errors';
-
-export const QuestionResponseFormat = {
-    type: 'json_object',
-    schema: {
-        type: 'object',
-        properties: {
-            questions: {
-                type: 'array',
-                items: {
-                    type: 'object',
-                    properties: {
-                        id: { type: 'number' },
-                        question: { type: 'string' },
-                        questionType: {
-                            type: 'string',
-                            enum: QuestionType,
-                            description: '질문 유형',
-                        },
-                    },
-                    required: ['id', 'question', 'questionType'],
-                },
-            },
-        },
-        required: ['questions'],
-    },
-};
+import { LLMService } from 'src/infra/llm/llm.service';
+import { Question } from 'src/common/types/question.type';
 
 @Injectable()
 export class MasterPortfoliosService {
-    private apiKey?: string;
-    private baseURL: string;
     constructor(
         @InjectRepository(Questions)
         private readonly questionsRepository: Repository<Questions>,
         @InjectRepository(MasterPortfolio)
         private readonly masterPortfolioRepository: Repository<MasterPortfolio>,
-        private readonly promptLoader: PromptLoader
-    ) {
-        this.apiKey = process.env.OPENROUTER_API_KEY;
-        this.baseURL = process.env.OPENROUTER_API_BASE_URL || 'https://openrouter.ai/api/v1';
-    }
+        private readonly llmService: LLMService
+    ) {}
 
     async createQuestions(projectId: number) {
         // 프로젝트 ID로 마스터 포트폴리오를 찾습니다.
@@ -58,51 +27,10 @@ export class MasterPortfoliosService {
         }
         const masterPortfolioId = masterPortfolio.id;
 
-        // 질문 생성 프롬프트를 로드합니다.
-        let questionPrompt: string;
-        try {
-            questionPrompt = await this.promptLoader.load('question.prompt.md');
-        } catch (e) {
-            throw new Error('Failed to load question generation prompt');
-        }
+        // LLM을 호출하여 질문을 생성합니다.
+        const questions: Array<Question> = await this.llmService.generateQuestions();
 
-        const headers = {
-            Authorization: `Bearer ${this.apiKey}`,
-            'Content-Type': 'application/json',
-        };
-        const completions = await fetch(`${this.baseURL}/chat/completions`, {
-            method: 'POST',
-            headers: headers,
-            body: JSON.stringify({
-                model:
-                    process.env.QUESTION_GENERATION_MODEL ||
-                    'google/gemini-2.5-flash-lite-preview-06-17',
-                messages: [
-                    {
-                        role: 'system',
-                        content: questionPrompt,
-                    },
-                ],
-                response_format: QuestionResponseFormat,
-            }),
-        });
-        if (!completions.ok) {
-            throw new Error(`Failed to create questions: ${completions.statusText}`);
-        }
-        const responseData = await completions.json();
-        let questions: Array<{
-            id: number;
-            question: string;
-            questionType: QuestionType;
-        }>;
-        try {
-            const responseJsonData = JSON.parse(responseData.choices[0].message.content);
-            questions = responseJsonData.questions;
-        } catch (e) {
-            // 별도 처리 필요
-            throw new InternalServerErrorException(`Failed to parse response: ${e.message}`);
-        }
-
+        // 생성된 질문들을 데이터베이스에 저장합니다.
         const questionEntities: Array<any> = [];
         for (const q of questions) {
             try {
@@ -123,7 +51,6 @@ export class MasterPortfoliosService {
                 throw new InternalServerErrorException(`Failed to create question: ${e.message}`);
             }
         }
-        console.log('저장된 질문들:', questionEntities);
         return questionEntities;
     }
